+++ bin/pubdoctor.dart
@@ -1,6 +1,6 @@
 import 'dart:io';
 
-import 'package:pubdoctor/src/cli/runner.dart';
+import 'package:pubdiagnose/src/cli/runner.dart';
 
 Future<void> main(List<String> args) async {
   final code = await runPubDoctor(args);