+++ test/pubdoctor_test.dart
@@ -1,4 +1,4 @@
-import 'package:pubdoctor/pubdoctor.dart';
+import 'package:pubdiagnose/pubdiagnose.dart';
 import 'package:test/test.dart';
 
 import 'helpers.dart';